% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDDworkshop2023-package.R
\docType{package}
\name{IDDworkshop2023-package}
\alias{IDDworkshop2023}
\alias{IDDworkshop2023-package}
\title{IDDworkshop2023: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CarmenTamayo/IDDworkshop2023}
  \item Report bugs at \url{https://github.com/CarmenTamayo/IDDworkshop2023/issues}
}

}
\author{
\strong{Maintainer}: Carmen Tamayo Cuartero \email{carmen.tamayo-cuartero@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-4184-2864}{ORCID}) [copyright holder]

}
\keyword{internal}
